requirements.txt
annotated-types==0.7.0
attrs==21.2.0
audioread==3.0.1
blinker==1.4
cachetools==5.5.1
certifi==2020.6.20
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.4.1
click==8.1.8
command-not-found==0.3
cryptography==44.0.0
dbus-python==1.2.18
decorator==5.1.1
distro==1.7.0
distro-info==1.1+ubuntu0.1
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
google-ai-generativelanguage==0.6.15
google-api-core==2.24.1
google-api-python-client==2.161.0
google-auth==2.38.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
google-cloud-texttospeech==2.24.0
google-generativeai==0.8.4
googleapis-common-protos==1.66.0
grpcio==1.70.0
grpcio-status==1.70.0
gTTS==2.5.4
httplib2==0.20.2
idna==3.3
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib-metadata==4.6.4
jeepney==0.7.1
joblib==1.4.2
jsonschema==3.2.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lazy_loader==0.4
librosa==0.10.2.post1
llvmlite==0.44.0
lxml==5.3.0
more-itertools==8.10.0
moviepy==2.1.2
msgpack==1.1.0
netifaces==0.11.0
numba==0.61.0
numpy==2.1.3
oauthlib==3.2.0
packaging==24.2
pdfminer.six==20231228
pdfplumber==0.11.5
pillow==10.4.0
platformdirs==4.3.6
pooch==1.8.2
proglog==0.1.10
proto-plus==1.26.0
protobuf==5.29.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydantic==2.10.6
pydantic_core==2.27.2
PyGObject==3.42.1
PyJWT==2.3.0
pyparsing==2.4.7
PyPDF2==3.0.1
pypdfium2==4.30.1
pyrsistent==0.18.1
python-apt==2.4.0+ubuntu2
python-dotenv==1.0.1
python-pptx==1.0.2
pyttsx3==2.98
PyYAML==5.4.1
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
scikit-learn==1.6.1
scipy==1.15.1
SecretStorage==3.3.1
six==1.16.0
soundfile==0.13.1
soxr==0.5.0.post1
systemd-python==234
texttable==1.6.4
threadpoolctl==3.5.0
tqdm==4.67.1
typing_extensions==4.12.2
ubuntu-advantage-tools==8001
ufw==0.36.1
unattended-upgrades==0.1
uritemplate==4.1.1
urllib3==1.26.5
wadllib==1.3.6
websocket-client==1.2.3
XlsxWriter==3.2.2
zipp==1.0.0
